Bug fixes for 3.23.23
myisam/mi_debug.c:
  ***MISSING WEAVE***
Docs/internals.texi:
  Added coding guidelines
Docs/manual.texi:
  Changelog update + Win32 -> Windows
client/mysql.cc:
  Changed --no-named-commands to be on by default
client/mysqlimport.c:
  Bug fix
include/config-win.h:
  Update of supported functions
include/global.h:
  Removed compiler warning
libmysql/libmysql.c:
  Fix for Ia64
myisam/ChangeLog:
  Changelog
myisam/Makefile.am:
  Added file mi_dbug.c
myisam/ft_stopwords.c:
  Fix for Ia64
myisam/mi_delete_table.c:
  Extra debugging
myisam/mi_rename.c:
  Extra debugging
myisam/mi_rnext.c:
  Fixed bug with MIN and concurrent insert
myisam/mi_rprev.c:
  Fixed bug with MAX and concurrent insert
myisam/mi_search.c:
  Fixed bug with DECIMAL/NUMERIC keys
myisam/myisamdef.h:
  Extra debugging
scripts/make_binary_distribution.sh:
  Added thread safe mysql library
sql/ha_heap.cc:
  Fix of HEAP bug with range keys
sql/ha_heap.h:
  Fix of HEAP bug with range keys
sql/handler.cc:
  Optimizing
sql/handler.h:
  Optimizing
sql/lock.cc:
  More DEBUG + fix of RENAME bug
sql/mini_client.cc:
  Fix for Ia64
sql/mysql_priv.h:
  Fix for name locks
sql/mysqld.cc:
  Shorter message if wrong options
sql/opt_range.cc:
  Added TODO
sql/sql_base.cc:
  Fix for DROP TABLE
sql/sql_parse.cc:
  Fix of permission checking for CHECK TABLE
sql/sql_select.cc:
  Fix of using LEFT JOIN with empty table
sql/table.h:
  Fix for name locks
tests/fork_test.pl:
  Fixed typo

+++ myisam/ft_stopwords.c
@@ -48,7 +48,7 @@ int ft_init_stopwords(const char **sws)
 
   for(;*sws;sws++)
   {
-    if( (sw.len=strlen(sw.pos=*sws)) < MIN_WORD_LEN) continue;
+    if( (sw.len= (uint) strlen(sw.pos=*sws)) < MIN_WORD_LEN) continue;
     if(!tree_insert(stopwords3, &sw, 0))
     {
       delete_tree(stopwords3);
